+++ qiskit/aqua/algorithms/single_sample/hhl/hhl.py
@@ -51,6 +51,18 @@ class HHL(QuantumAlgorithm):
             'id': 'hhl_schema',
             'type': 'object',
             'properties': {
+                'truncate_powerdim': {
+                    'type': 'boolean',
+                    'default': False
+                },
+                'truncate_hermitian': {
+                    'type': 'boolean',
+                    'default': False
+                },
+                'orig_size': {
+                    'type': ['integer', 'null'],
+                    'default': None
+                }
             },
             'additionalProperties': False
         },
@@ -82,11 +94,14 @@ class HHL(QuantumAlgorithm):
             self,
             matrix=None,
             vector=None,
+            truncate_powerdim=False,
+            truncate_hermitian=False,
             eigs=None,
             init_state=None,
             reciprocal=None,
             num_q=0,
-            num_a=0
+            num_a=0,
+            orig_size=None
     ):
         """
         Constructor.
@@ -94,16 +109,34 @@ class HHL(QuantumAlgorithm):
         Args:
             matrix (np.array): the input matrix of linear system of equations
             vector (np.array): the input vector of linear system of equations
+            truncate_powerdim (bool): flag indicating expansion to 2**n matrix to be truncated
+            truncate_hermitian (bool): flag indicating expansion to hermitian matrix to be truncated
             eigs (Eigenvalues): the eigenvalue estimation instance
             init_state (InitialState): the initial quantum state preparation
             reciprocal (Reciprocal): the eigenvalue reciprocal and controlled rotation instance
             num_q (int): number of qubits required for the matrix Operator instance
             num_a (int): number of ancillary qubits for Eigenvalues instance
+            orig_size (int): The original dimension of the problem (if truncate_powerdim)
         """
         super().__init__()
         super().validate(locals())
+        if matrix.shape[0] != matrix.shape[1]:
+            raise ValueError("Input matrix must be square!")
+        if matrix.shape[0] != len(vector):
+            raise ValueError("Input vector dimension does not match input "
+                             "matrix dimension!")
+        if not np.allclose(matrix, matrix.conj().T):
+            raise ValueError("Input matrix must be hermitian!")
+        if np.log2(matrix.shape[0]) % 1 != 0:
+            raise ValueError("Input matrix dimension must be 2**n!")
+        if truncate_powerdim and orig_size is None:
+            raise ValueError("Truncation to {} dimensions is not "
+                            "possible!".format(self._original_dimension))
+
         self._matrix = matrix
         self._vector = vector
+        self._truncate_powerdim = truncate_powerdim
+        self._truncate_hermitian = truncate_hermitian
         self._eigs = eigs
         self._init_state = init_state
         self._reciprocal = reciprocal
@@ -114,6 +147,7 @@ class HHL(QuantumAlgorithm):
         self._eigenvalue_register = None
         self._ancilla_register = None
         self._success_bit = None
+        self._original_dimension = orig_size
         self._ret = {}
 
     @classmethod
@@ -134,13 +168,32 @@ class HHL(QuantumAlgorithm):
         if not isinstance(vector, np.ndarray):
             vector = np.asarray(vector)
 
+        if matrix.shape[0] != matrix.shape[1]:
+            raise ValueError("Input matrix must be square!")
         if matrix.shape[0] != len(vector):
             raise ValueError("Input vector dimension does not match input "
                              "matrix dimension!")
-        if np.log2(matrix.shape[0]) % 1 != 0:
-            # TODO: extend vector and matrix for nonhermitian/non 2**n size
-            #  matrices and prune dimensions of HHL solution
-            raise ValueError("Matrix dimension must be 2**n!")
+
+        hhl_params = params.get(Pluggable.SECTION_KEY_ALGORITHM)
+        truncate_powerdim = hhl_params.get('truncate_powerdim')
+        truncate_hermitian = hhl_params.get('truncate_hermitian')
+        orig_size = hhl_params.get('orig_size')
+        if orig_size is None:
+            orig_size = len(vector)
+
+        is_powerdim = np.log2(matrix.shape[0]) % 1 == 0
+        if not is_powerdim:
+            logger.warning("Input matrix does not have dimension 2**n. It "
+                           "will be expanded automatically.")
+            matrix, vector = cls.expand_to_powerdim(matrix, vector)
+            truncate_powerdim = True
+
+        is_hermitian = np.allclose(matrix, matrix.conj().T)
+        if not is_hermitian:
+            logger.warning("Input matrix is not hermitian. It will be "
+                           "expanded to a hermitian matrix automatically.")
+            matrix, vector = cls.expand_to_hermitian(matrix, vector)
+            truncate_hermitian = True
 
         # Initialize eigenvalue finding module
         eigs_params = params.get(Pluggable.SECTION_KEY_EIGS)
@@ -163,7 +216,8 @@ class HHL(QuantumAlgorithm):
         reci = get_pluggable_class(PluggableType.RECIPROCAL,
                                    reciprocal_params['name']).init_params(params)
 
-        return cls(matrix, vector, eigs, init_state, reci, num_q, num_a)
+        return cls(matrix, vector, truncate_powerdim, truncate_hermitian, eigs,
+                   init_state, reci, num_q, num_a, orig_size)
 
     def construct_circuit(self, measurement=False):
         """Construct the HHL circuit.
@@ -191,9 +245,7 @@ class HHL(QuantumAlgorithm):
         s = self._reciprocal._anc
 
         # Inverse EigenvalueEstimation
-        qc += self._eigs.construct_inverse("circuit", self._eigs._circuit,
-                                           self._eigs._input_register,
-                                           self._eigs._output_register)
+        qc += self._eigs.construct_inverse("circuit", self._eigs._circuit)
 
         # Measurement of the ancilla qubit
         if measurement:
@@ -208,6 +260,80 @@ class HHL(QuantumAlgorithm):
         self._circuit = qc
         return qc
 
+    @staticmethod
+    def expand_to_powerdim(matrix, vector):
+        """ Expand a matrix to the next-larger 2**n dimensional matrix with
+        ones on the diagonal and zeros on the off-diagonal and expand the
+        vector with zeros accordingly.
+
+        Args:
+            matrix (np.array): the input matrix
+            vector (np.array): the input vector
+
+        Returns:
+            matrix (np.array): the expanded matrix
+            vector (np.array): the expanded vector
+        """
+        mat_dim = matrix.shape[0]
+        next_higher = int(np.ceil(np.log2(mat_dim)))
+        new_matrix = np.identity(2 ** next_higher)
+        new_matrix = np.array(new_matrix, dtype=complex)
+        new_matrix[:mat_dim, :mat_dim] = matrix[:, :]
+        matrix = new_matrix
+        new_vector = np.zeros((1, 2 ** next_higher))
+        new_vector[0, :vector.shape[0]] = vector
+        vector = new_vector.reshape(np.shape(new_vector)[1])
+        return matrix, vector
+
+    @staticmethod
+    def expand_to_hermitian(matrix, vector):
+        """ Expand a non-hermitian matrix A to a hermitian matrix by
+        [[0, A.H], [A, 0]] and expand vector b to [b.conj, b].
+
+        Args:
+            matrix (np.array): the input matrix
+            vector (np.array): the input vector
+
+        Returns:
+            matrix (np.array): the expanded matrix
+            vector (np.array): the expanded vector
+        """
+        #
+        half_dim = matrix.shape[0]
+        full_dim = 2 * half_dim
+        new_matrix = np.zeros([full_dim, full_dim])
+        new_matrix = np.array(new_matrix, dtype=complex)
+        new_matrix[0:half_dim, half_dim:full_dim] = matrix[:, :]
+        new_matrix[half_dim:full_dim, 0:half_dim] = matrix.conj().T[:, :]
+        matrix = new_matrix
+        new_vector = np.zeros((1, full_dim))
+        new_vector = np.array(new_vector, dtype=complex)
+        new_vector[0, :vector.shape[0]] = vector.conj()
+        new_vector[0, vector.shape[0]:] = vector
+        vector = new_vector.reshape(np.shape(new_vector)[1])
+        return matrix, vector
+
+    def _resize_vector(self, vec):
+        if self._truncate_hermitian:
+            half_dim = int(vec.shape[0] / 2)
+            vec = vec[:half_dim]
+        if self._truncate_powerdim:
+            vec = vec[:self._original_dimension]
+        return vec
+
+    def _resize_matrix(self, matrix):
+        if self._truncate_hermitian:
+            full_dim = matrix.shape[0]
+            half_dim = int(full_dim / 2)
+            new_matrix = np.ndarray(shape=(half_dim, half_dim), dtype=complex)
+            new_matrix[:, :] = matrix[0:half_dim, half_dim:full_dim]
+            matrix = new_matrix
+        if self._truncate_powerdim:
+            new_matrix = np.ndarray(shape=(self._original_dimension, self._original_dimension), dtype=complex)
+            new_matrix[:, :] = matrix[:self._original_dimension, :self._original_dimension]
+            matrix = new_matrix
+        return matrix
+
     def _statevector_simulation(self):
         """The statevector simulation.
 
@@ -218,7 +344,8 @@ class HHL(QuantumAlgorithm):
         sv = np.asarray(res.get_statevector(self._circuit))
         # Extract solution vector from statevector
         vec = self._reciprocal.sv_to_resvec(sv, self._num_q)
-        self._ret['probability_result'] = np.real(vec.dot(vec.conj()))
+        # remove added dimensions
+        self._ret['probability_result'] = np.real(self._resize_vector(vec).dot(self._resize_vector(vec).conj()))
         vec = vec/np.linalg.norm(vec)
         self._hhl_results(vec)
 
@@ -250,6 +377,7 @@ class HHL(QuantumAlgorithm):
                 else:
                     f += v
             probs.append(s/(f+s))
+        probs = self._resize_vector(probs)
         self._ret["probability_result"] = np.real(probs)
 
         # Filtering the tomo data for valid results with ancillary measured
@@ -287,12 +415,15 @@ class HHL(QuantumAlgorithm):
         return new_results
 
     def _hhl_results(self, vec):
-        self._ret["output"] = vec
+        res_vec = self._resize_vector(vec)
+        in_vec = self._resize_vector(self._vector)
+        matrix = self._resize_matrix(self._matrix)
+        self._ret["output"] = res_vec
         # Rescaling the output vector to the real solution vector
-        tmp_vec = self._matrix.dot(vec)
-        f1 = np.linalg.norm(self._vector)/np.linalg.norm(tmp_vec)
-        f2 = sum(np.angle(self._vector*tmp_vec.conj()-1+1))/self._num_q  # "-1+1" to fix angle error for -0.-0.j
-        self._ret["solution"] = f1*vec*np.exp(-1j*f2)
+        tmp_vec = matrix.dot(res_vec)
+        f1 = np.linalg.norm(in_vec)/np.linalg.norm(tmp_vec)
+        f2 = sum(np.angle(in_vec*tmp_vec.conj()-1+1))/(np.log2(matrix.shape[0])) # "-1+1" to fix angle error for -0.-0.j
+        self._ret["solution"] = f1*res_vec*np.exp(-1j*f2)
 
     def _run(self):
         if self._quantum_instance.is_statevector:
@@ -302,7 +433,7 @@ class HHL(QuantumAlgorithm):
             self.construct_circuit(measurement=False)
             self._state_tomography()
         # Adding a bit of general result information
-        self._ret["matrix"] = self._matrix
-        self._ret["vector"] = self._vector
+        self._ret["matrix"] = self._resize_matrix(self._matrix)
+        self._ret["vector"] = self._resize_vector(self._vector)
         self._ret["circuit_info"] = circuit_to_dag(self._circuit).properties()
         return self._ret